vga_buffer: Implement Writer type
The writer will always write to the last line and shift lines up when
a line is full. The new_line() function will be implemented soon.

Rust strings are UTF-8 by default, so they might contain bytes that are
not supported by the VGA text buffer. We use a match to differentiate
printable ASCII bytes (a newline or anything in between a space character
and a ~ character) and unprintable bytes. For unprintable bytes, we print
a ■ character, which has the hex code 0xfe on the VGA hardware.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>