vga_buffer: Use volatile to ensure writes to the buffer
The problem is that we only write to the Buffer and never read from it
again. The compiler doesn’t know that we really access VGA buffer memory
(instead of normal RAM) and knows nothing about the side effect that
some characters appear on the screen. So it might decide that these
writes are unnecessary and can be omitted.

To avoid this erroneous optimization, we need to specify these writes
as volatile. This tells the compiler that the write has side effects
and should not be optimized away.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>