vga_buffer: Use a global Writer as an interface for other modules
Use the lazy_static crate to create our static WRITER. However, this WRITER
is pretty useless since it is immutable (all write methods take &mut self).

To get synchronizzed interior mutability, use a spinlock to synchronize
writes to our WRITER.

With this, we encoded the required conditions in the type system and are
able to provide a safe Writer interface to the outside.

Signed-off-by: Piyush Kumar <piyushkumg@gmail.com>